python_base/lesson_013/requirements.txt
Pillow==8.1.0
pyspark==2.2.3
